refactor: extract components from analytics, admin, trade-history pages
- analytics/metric-card.tsx
- admin/stat-card.tsx, document-row.tsx, agent-card.tsx, agent-constants.ts
- trade-history/trade-detail-modal.tsx
- All pages updated to import from components/
