+++ apps/web/src/components/admin/index.ts
@@ -1,0 +1,8 @@
+export { StatCard } from './stat-card';
+export { DocumentRow } from './document-row';
+export { AgentCard } from './agent-card';
+export {
+  AGENT_ICON_MAP,
+  AGENT_COLOR_MAP,
+  AGENT_NAMES,
+} from './agent-constants';
