+++ apps/web/src/components/analytics/index.ts
@@ -1,0 +1,1 @@
+export { MetricCard } from './metric-card';
