+++ apps/web/src/components/trade-history/index.ts
@@ -1,0 +1,1 @@
+export { TradeDetailModal } from './trade-detail-modal';
