+++ apps/web/src/pages/admin/stats.tsx
@@ -8,7 +8,7 @@ import {
   Clock,
   Loader2,
 } from 'lucide-react';
-import { Button, InfoTooltip } from '@crypto-trader/ui';
+import { Button } from '@crypto-trader/ui';
 import { cn } from '../../lib/utils';
 import { useGSAP } from '@gsap/react';
 import gsap from 'gsap';
@@ -19,31 +19,7 @@ import {
   useKillSwitch,
 } from '../../hooks/use-admin';
 import { useTranslation } from 'react-i18next';
-
-function StatCard({
-  label,
-  value,
-  icon: Icon,
-  color,
-  tooltip,
-}: {
-  label: string;
-  value: string | number;
-  icon: typeof Activity;
-  color?: string;
-  tooltip?: string;
-}) {
-  return (
-    <div className="stat-card rounded-xl border border-border bg-card p-5">
-      <div className="mb-1 flex items-center gap-1.5 text-sm text-muted-foreground">
-        <Icon className={cn('h-4 w-4', color ?? 'text-primary')} />
-        {label}
-        {tooltip && <InfoTooltip text={tooltip} />}
-      </div>
-      <div className="text-2xl font-bold">{value}</div>
-    </div>
-  );
-}
+import { StatCard } from '../../components/admin';
 
 export function AdminStatsPage() {
   const { t } = useTranslation();
